moved and refactor route and component names
