+++ src/components/GamePage/DisplayAndQuestion/DisplayAndQuestion.js
@@ -7,9 +7,7 @@ export default class DisplayAndQuestion extends Component {
     static contextType = GameSessionContext;
 
     checkForResponseOrQuestion = () => {
-        //console.log(this.context.inResponseState);
         if(this.context.inResponseState){
-            //console.log("were in response state");
             return <ResponseDisplay />
         } else {
             return <QuestionDisplay />
