+++ src/components/GamePage/ResponseDisplay/ResponseDisplay.js
@@ -9,12 +9,9 @@ export default class ResponseDisplay extends Component {
         questionLength: 1,
     }
     progressToQuestion = () => {
-        //console.log(this.context.response)
         if(this.context.response.game_ending){
-            console.log('Set the game to lose')
             this.context.setGameLost();
         }else if (this.context.lastQuestion) {
-            console.log('Set the game to complete');
             this.context.setGameComplete();
         } else {
             this.context.grabNewData();
@@ -23,14 +20,12 @@ export default class ResponseDisplay extends Component {
     }
 
     renderResponse = (responseText) => {
-        //console.log(responseText);
         const textArr = responseText.split('');
         let renderText = "";
 
         this.renderInterval = setInterval(() => {
             if (this.state.counter < textArr.length) {
                 renderText = this.state.responseText + textArr[this.state.counter];
-                //console.log(this.state.counter);
                 this.setState({
                     responseText: renderText,
                     counter: this.state.counter + 1,
@@ -55,7 +50,6 @@ export default class ResponseDisplay extends Component {
         let responseText = "";
         //Let response load in
         if (this.context.response.response) {
-            //console.log(this.context.response);
             responseText = this.context.response.response
             //Sets the newlines
             responseText = responseText.replace(/\\n/g, '\n');
