+++ src/components/GamePage/StartSection/StartSection.css
