+++ src/routes/CharSelectPage/CharSelectPage.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import CharSelectSection from '../../../components/GamePage/CharSelect/CharSelect'
+import CharSelectSection from '../../components/GamePage/CharSelect/CharSelect'
 export default class CharSelectPage extends Component {
     render() {
         return (
