+++ src/routes/PlayGamePage/PlayGamePage.js
@@ -1,16 +1,15 @@
 import React, { Component } from 'react';
 import { Route } from 'react-router-dom';
-import PlayScreenMain from '../../../components/GamePage/PlayScreenMain/PlayScreenMain';
-import PlayScreenInventory from '../../../components/GamePage/PlayScreenInventory/PlayScreenInventory';
-import NoGameSession from '../../../components/LandingPage/NoGameSession/NoGameSession';
-import GameContext from '../../../contexts/GameContext';
+import PlayScreenMain from '../../components/GamePage/PlayScreenMain/PlayScreenMain';
+import PlayScreenInventory from '../../components/GamePage/PlayScreenInventory/PlayScreenInventory';
+import NoGameSession from '../../components/LandingPage/NoGameSession/NoGameSession';
+import GameContext from '../../contexts/GameContext';
 
 export default class ClassName extends Component {
 
     static contextType = GameContext;
 
     makeGameRoutes = () => {
-        //console.log(this.context.characterSelected);
         if(this.context.gameInSession){
             return (
                 <>
